cyra/requirements.txt
# Core AI and API dependencies
openai>=1.12.0
azure-cognitiveservices-speech>=1.34.1
azure-identity>=1.15.0

# Web framework and API
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=12.0

# Database and authentication
sqlalchemy>=2.0.23
alembic>=1.13.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Security tools
python-nmap>=0.7.1
cryptography>=41.0.8
requests>=2.31.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.2
pydantic-settings>=2.1.0

# Audio processing
pyaudio>=0.2.11
wave>=0.0.2

# Utility libraries
rich>=13.7.0
typer>=0.9.0
asyncio-mqtt>=0.13.0

# Development dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
